GraphAltair/src/requirements.txt
altair==5.2.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.1.1
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fonttools==4.47.2
h11==0.14.0
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
jsonpickle==3.0.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
neo4j==5.16.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
nx-altair==0.1.6
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
portpicker==1.6.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyvis==0.3.2
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
scipy==1.12.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
toolz==0.12.1
tornado==6.4
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vega-datasets==0.9.0
vl-convert-python==1.2.3
wcwidth==0.2.13
wsproto==1.2.0
